+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Restruct `JsonRpcNotification` class.
 - `Tool` constructor validates `name`.
 - Rename `Tool.parameters` to `Tool.inputSchema`.
+- Deprecate `ProtocolHandler.sendNotificationToClient` method.
 
 ## 0.5.0